+++ app/javascript/packs/application.js
@@ -14,7 +14,7 @@
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
 
-import App from "../app.vue";
+import App from "../App.vue";
 import Vue from "vue";
 
 import VueRouter from "vue-router";
