+++ app/javascript/packs/components/RegisteredTags.vue
