+++ app/javascript/packs/components/shared/TheFooter.vue
